requirements.txt
Django==5.2.13
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
python-dotenv==1.2.2
Pillow==12.2.0
django-allauth==65.16.1
requests==2.33.1
# MySQL 연결 시 추가: mysqlclient==2.2.x
